requirements.txt
asgiref==3.7.2
astroid==2.15.6
dill==0.3.7
Django==4.2.5
djangorestframework==3.14.0
Faker==19.6.2
flake8==6.1.0
gunicorn==21.2.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
packaging==23.1
platformdirs==3.10.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pyflakes==3.1.0
pylint==2.17.5
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.8.0
wrapt==1.15.0
